+++ CodexBarMobile/CodexBarMobile/CodexBarMobileApp.swift
@@ -26,6 +26,7 @@ struct CodexBarMobileApp: App {
             defaults.removeObject(forKey: MobileSettingsKeys.usagePercentDisplayMode)
             defaults.removeObject(forKey: MobileSettingsKeys.showRemainingUsage)
             defaults.removeObject(forKey: MobileSettingsKeys.freeSelectedProviderID)
+            defaults.removeObject(forKey: MobileSettingsKeys.freeSelectedProviderLockedUntil)
             defaults.removeObject(forKey: ProEntitlementCacheStore.key)
             defaults.removeObject(forKey: "onboardingSeenVersion")
             defaults.removeObject(forKey: MobileSettingsKeys.appearanceMode)